move enums.py to api dir
